Refactor move CPBKGB22::CHECKING => CPBKGB22::Personal::Current .
'CHECKING' was used, as that's the account type in OFX. However, in the UK,
this type of account is usually called 'Current'. In addition, The Co-operative
Bank has a top-level split of account types (same BIC, different sort codes),
and the input formats written so far are both for 'Personal' accounts.

Rename throughout, to make clearer for users, and to anticipate the addition
of Business input formats [#5].