[#2] Fix CPBKGB22::Personal::CreditCard::V_2015_05_27 for 4-column header.
The original 2015-05-27 format had 5 columns in the header. A slight change to
the format changed this to 4 columns. It doesn't seem worth creating a whole new
input parser, as everything else seems to still work.

This is so brittle. Oh for HTML with ids or even a range of classes... We could
match on text instead, but that doesn't make me feel confident, either.